// backend/src/employer/types/employer.types.ts
export interface EmployerCurrentUser {
  userAccountId: number;
  userRole: string;
}

export interface PaginatedResponse<T> {
  data: T[];
  meta: {
    page: number;
    limit: number;
    total: number;
    totalPages: number;
  };
}

export interface CompanyIdentity {
  companyId: number;
  userAccountId: number;
}